:recycle: refactor(service) : 불필요한 패키지 삭제
